% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function gets passed a specific state and year, then gets associated data.
If the state exists, it filters the data per state and plots accidents for that state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number indicating a state}

\item{year}{A number indicating the year}
}
\value{
This function creates a graph for a specific state and year
}
\description{
This function gets passed a specific state and year, then gets associated data.
If the state exists, it filters the data per state and plots accidents for that state.
}
\note{
This function returns an error if the state number is not valid
}
\examples{
\dontrun{
fars_map_state(state.num=1,year=2015)
}

}
