% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{fars functions for Coursera R package course}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string indicating the name of csv file}
}
\value{
This function returns a table with data from the loaded file
}
\description{
The fars_read function imports a file, if it exists, in csv format and
converts it to a table. If file does not exist it returns an error.
}
\note{
Function will return error if file does not exist
}
\examples{
\dontrun{
fars_read(filename="accident_2015.csv.bz2")
}

}
