% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This function makes data frames with MONTH and year columns}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector representing years}
}
\value{
This function returns a list of tables corresponding to list of years.
}
\description{
When passing the function a list of valid years as input, it gets associated files,
For each file it retrieves the corresponding data table and selects the MONTH and year columns.
}
\note{
Function will return an error if year is not valid
}
\examples{
\dontrun{
fars_read_years(years=c(2013,2014,2015))
}

}
