% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This function receives a list of years as an input and gets data tables
with \code{fars_read_years} Then it summarizes the results per year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector representing years}
}
\value{
This function returns a table with summarized results for each year.
}
\description{
This function receives a list of years as an input and gets data tables
with \code{fars_read_years} Then it summarizes the results per year.
}
