% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{This function creates the file name when given \code{year} as input}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number indicating the year}
}
\value{
This function returns a character string with coding "accident_year.csv.bz2"
}
\description{
This function creates the file name when given \code{year} as input
}
\examples{
\dontrun{
make_filename(2015)
}
}
